[Twinkle#67#2] Update data types in models

+++ twinkle-app/twinkle.dal.abstractions/src/main/java/com/github/donkeyrit/twinkle/dal/interfaces/CarRepository.java
@@ -2,12 +2,12 @@ package com.github.donkeyrit.twinkle.dal.interfaces;
 
 import com.github.donkeyrit.twinkle.dal.common.repositories.FilterableRepository;
 import com.github.donkeyrit.twinkle.dal.common.specifications.PagedSpecification;
+import com.github.donkeyrit.twinkle.dal.specifications.CarQuerySpecification;
 import com.github.donkeyrit.twinkle.dal.common.models.Page;
 import com.github.donkeyrit.twinkle.dal.models.Car;
-import com.github.donkeyrit.twinkle.dal.specifications.CarQuerySpecification;
 
 public interface CarRepository extends FilterableRepository<Car, CarQuerySpecification> 
 {
 	Page<Car> getPagedResult(PagedSpecification<Car> filter);
-	int getMaxPrice();
+	Double getMaxPrice();
 }
